# quarkus-workshop-super-heroes/super-heroes/rest-hero/openapi.yaml
---
openapi: 3.0.3
info:
    title: rest-hero API
    version: 1.0.0-SNAPSHOT
paths:
    /api/heroes:
        get:
            summary: Returns all the heroes from the database
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Hero'
                "204":
                    description: No heroes
        put:
            summary: Updates an exiting  hero
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Hero'
                required: true
            responses:
                "200":
                    description: The updated hero
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Hero'
        post:
            summary: Creates a valid hero
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Hero'
                required: true
            responses:
                "201":
                    description: The URI of the created hero
                    content:
                        application/json:
                            schema:
                                format: uri
                                type: string
    /api/heroes/hello:
        get:
            responses:
                "200":
                    description: OK
                    content:
                        text/plain:
                            schema:
                                type: string
    /api/heroes/random:
        get:
            summary: Returns a random hero
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Hero'
    /api/heroes/{id}:
        get:
            summary: Returns a hero for a given identifier
            parameters:
                - name: id
                  in: path
                  description: Hero identifier
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Hero'
                "204":
                    description: The hero is not found for a given identifier
        delete:
            summary: Deletes an exiting hero
            parameters:
                - name: id
                  in: path
                  description: Hero identifier
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                "204": {}
components:
    schemas:
        Hero:
            required:
                - level
                - name
            type: object
            properties:
                id:
                    format: int64
                    type: integer
                level:
                    format: int32
                    minimum: 1
                    type: integer
                    nullable: false
                name:
                    maxLength: 50
                    minLength: 3
                    type: string
                    nullable: false
                otherName:
                    type: string
                picture:
                    type: string
                powers:
                    type: string